requirements.txt
Django==5.1.5
djangorestframework==3.15.2
firebase-admin==6.6.0
python-dotenv==1.0.1
django-cors-headers==4.6.0
requests==2.32.3
